requirements.txt
azure-common==1.1.28
azure-core==1.32.0
azure-cosmos==4.7.0
azure-functions==1.21.3
azure-identity==1.19.0
azure-keyvault==4.2.0
azure-keyvault-certificates==4.9.0
azure-keyvault-keys==4.10.0
azure-keyvault-secrets==4.9.0
azure-mgmt-core==1.5.0
azure-mgmt-cosmosdb==9.6.0
azure-storage-queue==12.12.0
boto3==1.35.53
botocore==1.35.53
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
idna==3.10
isodate==0.7.2
jmespath==1.0.1
msal==1.31.0
msal-extensions==1.2.0
portalocker==2.10.1
pycparser==2.22
PyJWT==2.9.0
pyodbc==5.2.0
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.10.3
six==1.16.0
sqlglot==25.28.0
typing_extensions==4.12.2
urllib3==1.26.20
